requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0